경북대 BE_박준석_4주차 과제 (3단계, 4단계) (#411)
* docs(README.md): add functional requirements for step3

* feat(Option.java): add option quantity deduction

* feat: change so that options can be added on admin page

* feat: change option to delete on admin page

* feat: change so that options can be edited on admin page

* test(OptionRepositoryTest.java): add option quantity deduction in option repository test

* chore(InsufficientQuantityException.java): add remaining quantity in insufficient quantity exception

* fix: change product to have at least one option

* fix: apply optimistic lock to option

+++ src/main/java/gift/service/OptionService.java
@@ -3,12 +3,14 @@ package gift.service;
 import gift.domain.Option;
 import gift.domain.Product;
 import gift.dto.OptionDTO;
+import gift.exception.NoOptionsForProductException;
 import gift.exception.NoSuchOptionException;
 import gift.exception.NoSuchProductException;
 import gift.repository.OptionRepository;
 import gift.repository.ProductRepository;
 import java.util.List;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 public class OptionService {
@@ -45,10 +47,23 @@ public class OptionService {
         return optionRepository.save(updatedOption).toDTO();
     }
 
-    public OptionDTO deleteOption(long id) {
+    public OptionDTO deleteOption(long productId, long id) {
+        Product product = productRepository.findById(productId)
+            .orElseThrow(NoSuchProductException::new);
+        if (product.getOptions().size() == 1) {
+            throw new NoOptionsForProductException();
+        }
         Option deletedOption = optionRepository.findById(id)
             .orElseThrow(NoSuchOptionException::new);
         optionRepository.delete(deletedOption);
         return deletedOption.toDTO();
+    }
+
+    @Transactional
+    public OptionDTO buyOption(long id, int quantity) {
+        Option option = optionRepository.findById(id)
+            .orElseThrow(NoSuchOptionException::new);
+        option.subtract(quantity);
+        return optionRepository.save(option).toDTO();
     }
 }
